refactor: break away util helpers from validation helpers
